requirements.txt
uvicorn
fastapi
langchain-google-genai
sentence-transformers
PyMuPDF
python-dotenv
transformers